Removed EngineSerializer trait.
