+++ js/src/lib.rs
@@ -4,7 +4,6 @@ use adblock::lists::{
 use adblock::resources::resource_assembler::assemble_web_accessible_resources;
 use adblock::resources::Resource;
 use adblock::Engine as EngineInternal;
-use adblock::EngineSerializer as EngineSerializerInternal;
 use neon::prelude::*;
 use neon::types::buffer::TypedArray as _;
 use serde::{Deserialize, Serialize};
