+++ src/engine.rs
@@ -244,6 +244,33 @@ impl Engine {
     pub fn get_regex_debug_info(&self) -> crate::regex_manager::RegexDebugInfo {
         self.blocker.get_regex_debug_info()
     }
+
+    /// Serializes the `Engine` into a binary format so that it can be quickly reloaded later.
+    pub fn serialize(&self) -> Result<Vec<u8>, crate::data_format::SerializationError> {
+        use crate::data_format::SerializeFormat;
+
+        let serialize_format = SerializeFormat::build(&self.blocker, &self.cosmetic_cache);
+
+        let result = serialize_format.serialize()?;
+        Ok(result)
+    }
+
+    /// Deserialize the `Engine` from the binary format generated by `Engine::serialize`. The
+    /// method will automatically select the correct deserialization implementation.
+    pub fn deserialize(
+        &mut self,
+        serialized: &[u8],
+    ) -> Result<(), crate::data_format::DeserializationError> {
+        use crate::data_format::DeserializeFormat;
+        let current_tags = self.blocker.tags_enabled();
+        let deserialize_format = DeserializeFormat::deserialize(serialized)?;
+        let (blocker, cosmetic_cache) = deserialize_format.build();
+        self.blocker = blocker;
+        self.blocker
+            .use_tags(&current_tags.iter().map(|s| &**s).collect::<Vec<_>>());
+        self.cosmetic_cache = cosmetic_cache;
+        Ok(())
+    }
 }
 
 /// Static assertions for `Engine: Send + Sync` traits.
