+++ src/lib.rs
@@ -22,7 +22,6 @@ pub mod content_blocking;
 pub mod cosmetic_filter_cache;
 mod data_format;
 mod engine;
-mod engine_serializer;
 pub mod filters;
 pub mod lists;
 mod network_filter_list;
@@ -37,7 +36,6 @@ pub mod utils;
 
 #[doc(inline)]
 pub use engine::Engine;
-pub use engine_serializer::EngineSerializer;
 #[doc(inline)]
 pub use lists::FilterSet;
