+++ tests/legacy_harness.rs
@@ -316,7 +316,7 @@ mod legacy_test_filters {
 
 mod legacy_check_match {
     use adblock::request::Request;
-    use adblock::{Engine, EngineSerializer};
+    use adblock::Engine;
 
     fn check_match<'a>(
         rules: &[&'a str],
@@ -855,7 +855,6 @@ mod legacy_misc_tests {
     use adblock::filters::network::NetworkFilter;
     use adblock::request::Request;
     use adblock::Engine;
-    use adblock::EngineSerializer;
 
     #[test]
     fn demo_app() {
